Revert "Sameer"
