+++ client/App.jsx
@@ -48,8 +48,6 @@ function App() {
     return (
     <Router>
         <Navbar />
-        {/* <Canvas /> */}
-        <Navbar />
         <div className="App">
             <Navbar isLoggedIn={cookiePresent}/>
             <Routes>
