+++ package.json
@@ -43,7 +43,6 @@
     "@reduxjs/toolkit": "^2.2.3",
     "aws-sdk": "^2.1601.0",
     "axios": "^1.6.8",
-    "babel-node": "^0.0.1-security",
     "bcryptjs": "^2.4.3",
     "concurrently": "^8.2.2",
     "cookie-parser": "^1.4.6",
@@ -63,7 +62,6 @@
     "passport-google-oauth2": "^0.2.0",
     "react": "^18.2.0",
     "react-color": "^2.19.3",
-    "react-colorful": "^5.6.1",
     "react-dom": "^18.2.0",
     "react-redux": "^9.1.0",
     "react-router-dom": "^6.22.3",
